backend/requirements.txt
flask>=3.0,<4
flask-cors>=4.0,<5
firebase-admin>=6.2,<7
python-dotenv>=1.0,<2
gunicorn>=21.2,<22

torch>=2.2,<3
torchvision>=0.17,<1
torchaudio>=2.2,<3
numpy>=1.26,<3
scikit-learn>=1.3,<2
pandas>=2.1,<3
Pillow>=10,<11
onnx>=1.15,<2
onnxruntime>=1.17,<2
